Fixes the dtype of the environment origins in Scene
# Description

Fixes the dtype of the environment origins when obtained from cloner to be `float32` instead of the `float64`.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/scene/interactive_scene.py
@@ -135,7 +135,7 @@ class InteractiveScene:
                 replicate_physics=False,
                 copy_from_source=True,
             )
-        self._default_env_origins = torch.tensor(env_origins, device=self.device)
+        self._default_env_origins = torch.tensor(env_origins, device=self.device, dtype=torch.float32)
         # add entities from config
         self._add_entities_from_cfg()
         # replicate physics if we have more than one environment
